feat/#336: 로그인 상태에 따른 랜딩페이지 조건 분기
